+++ hazelcast-client/src/test/java/com/hazelcast/client/scheduledexecutor/ClientScheduledExecutorServiceSlowTest.java
@@ -1,4 +1,21 @@
 /*
+ * Copyright (c) 2008-2017, Hazelcast, Inc. All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ */
+
+/*
  * Copyright (c) 2008-2016, Hazelcast, Inc. All Rights Reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
@@ -20,16 +37,26 @@ import com.hazelcast.client.test.TestHazelcastFactory;
 import com.hazelcast.config.Config;
 import com.hazelcast.core.HazelcastInstance;
 import com.hazelcast.scheduledexecutor.IScheduledExecutorService;
-import com.hazelcast.scheduledexecutor.ScheduledExecutorServiceTest;
+import com.hazelcast.scheduledexecutor.ScheduledExecutorServiceSlowTest;
 import com.hazelcast.test.HazelcastParallelClassRunner;
 import com.hazelcast.test.annotation.ParallelTest;
-import com.hazelcast.test.annotation.QuickTest;
+import com.hazelcast.test.annotation.SlowTest;
+import org.junit.After;
 import org.junit.experimental.categories.Category;
 import org.junit.runner.RunWith;
 
 @RunWith(HazelcastParallelClassRunner.class)
-@Category({QuickTest.class, ParallelTest.class})
-public class ClientScheduledExecutorServiceTest extends ScheduledExecutorServiceTest {
+@Category({SlowTest.class, ParallelTest.class})
+public class ClientScheduledExecutorServiceSlowTest extends ScheduledExecutorServiceSlowTest {
+
+    private TestHazelcastFactory factory;
+
+    @After
+    public void teardown() {
+        if (factory != null) {
+            factory.terminateAll();
+        }
+    }
 
     @Override
     public HazelcastInstance[] createClusterWithCount(int count) {
@@ -39,18 +66,6 @@ public class ClientScheduledExecutorServiceTest extends ScheduledExecutorService
 
     @Override
     public IScheduledExecutorService getScheduledExecutor(HazelcastInstance[] instances, String name) {
-        return ((TestHazelcastFactory) factory).newHazelcastClient().getScheduledExecutorService(name);
-    }
-
-    @Override
-    public void getErroneous()
-            throws InterruptedException {
-        // Ignore - pending client change to support the ExecutionException wrapper
-    }
-
-    @Override
-    public void getErroneous_durable()
-            throws InterruptedException {
-        // Ignore - pending client change to support the ExecutionException wrapper
+        return factory.newHazelcastClient().getScheduledExecutorService(name);
     }
 }
