+++ hazelcast/src/test/java/com/hazelcast/scheduledexecutor/ScheduledHazelcastInstanceAwareTest.java
@@ -18,10 +18,7 @@
 package com.hazelcast.scheduledexecutor;
 
 import com.hazelcast.core.HazelcastInstance;
-import com.hazelcast.core.HazelcastInstanceAware;
 import com.hazelcast.test.HazelcastParallelClassRunner;
-import com.hazelcast.test.HazelcastTestSupport;
-import com.hazelcast.test.TestHazelcastInstanceFactory;
 import com.hazelcast.test.annotation.ParallelTest;
 import com.hazelcast.test.annotation.QuickTest;
 import org.junit.Before;
@@ -29,8 +26,6 @@ import org.junit.Test;
 import org.junit.experimental.categories.Category;
 import org.junit.runner.RunWith;
 
-import java.io.Serializable;
-import java.util.concurrent.Callable;
 import java.util.concurrent.ExecutionException;
 
 import static java.util.concurrent.TimeUnit.MILLISECONDS;
@@ -38,17 +33,15 @@ import static org.junit.Assert.assertTrue;
 
 @RunWith(HazelcastParallelClassRunner.class)
 @Category({QuickTest.class, ParallelTest.class})
-public class ScheduledHazelcastInstanceAwareTest extends HazelcastTestSupport {
+public class ScheduledHazelcastInstanceAwareTest extends ScheduledExecutorServiceTestSupport {
 
-    private TestHazelcastInstanceFactory factory;
     private HazelcastInstance[] members;
     protected IScheduledExecutorService scheduledExecutorService;
 
     @Before
     public void setup() {
-        factory = createHazelcastInstanceFactory(2);
-        members = factory.newInstances();
-        scheduledExecutorService = members[0].getScheduledExecutorService(randomName());
+        members = createClusterWithCount(2);
+        scheduledExecutorService = getScheduledExecutor(members, randomName());
     }
 
     @Test
@@ -65,30 +58,5 @@ public class ScheduledHazelcastInstanceAwareTest extends HazelcastTestSupport {
         IScheduledFuture<Boolean> injected = scheduledExecutorService.schedule(
                 new HazelcastInstanceAwareRunnable(randomNameOwnedBy(members[1])),200, MILLISECONDS);
         assertTrue("HazelcastInstance should have been injected", injected.get());
-    }
-
-    public static class HazelcastInstanceAwareRunnable implements Callable<Boolean>, HazelcastInstanceAware, Serializable,
-                                                                  NamedTask {
-        private transient volatile HazelcastInstance instance;
-        private final String name;
-
-        public HazelcastInstanceAwareRunnable(String name) {
-            this.name = name;
-        }
-
-        @Override
-        public void setHazelcastInstance(final HazelcastInstance instance) {
-            this.instance = instance;
-        }
-
-        @Override
-        public String getName() {
-            return name;
-        }
-
-        @Override
-        public Boolean call() {
-            return (instance != null);
-        }
     }
 }
